+++ traffic_ops/testing/api/v3/federation_users_test.go
@@ -20,7 +20,7 @@ import (
 )
 
 func TestFederationUsers(t *testing.T) {
-	WithObjs(t, []TCObj{CDNs, Types, Tenants, Users, Parameters, Profiles, Statuses, Divisions, Regions, PhysLocations, CacheGroups, DeliveryServices, CDNFederations, FederationUsers}, func() {
+	WithObjs(t, []TCObj{CDNs, Types, Tenants, Users, Parameters, Profiles, Statuses, Divisions, Regions, PhysLocations, CacheGroups, Topologies, DeliveryServices, CDNFederations, FederationUsers}, func() {
 		CreateTestInvalidFederationUsers(t)
 		GetTestInvalidFederationIDUsers(t)
 	})