+++ traffic_portal/app/src/modules/private/deliveryServices/edit/index.js
@@ -50,6 +50,9 @@ module.exports = angular.module('trafficPortal.private.deliveryServices.edit', [
                             origin: function($stateParams, originService) {
                                 return originService.getOrigins({ deliveryservice: $stateParams.deliveryServiceId, primary: true })
                             },
+                            topologies: function(topologyService) {
+                                return topologyService.getTopologies();
+                            },
                             type: function($stateParams) {
                                 return $stateParams.type;
                             },